+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Changes in the upcoming versions.
 - Add canister id input in the dummy relying party.
 - Upgrade `@dfinity/verifiable-credentials` used in the dummy relying party to the latest published version.
 - Port vc-util from II to this repository and rename it to `ic-verifiable-credentials`.
+- Do not publish to NPM the JS library if there were no changes.
 - Publish the dummy issuer and dummy relying party wasm modules and interface files as release artifacts so that they can be easily pulled into other projects.
 
 # 2024.06.10